SimilarityMeasureUsingCLIP-main/SimilarityMeasureUsingCLIP-main/requirements.txt
torch>=2.0.0
open_clip_torch>=2.24.0
pandas>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-image>=0.22.0
matplotlib>=3.8.0
